pyproject.toml
[tool.poetry]
name = "palettesnap"
version = "0.0.1"
description = "A colorscheme generator"
authors = ["EmperorEntropy"]
readme = "README.md"
license = "MIT"
repository = "https://github.com/EmperorEntropy/PaletteSnap"

[tool.poetry.scripts]
palsnap = "palettesnap.cli:app"

[tool.poetry.dependencies]
python = "~3.12"
toml = "^0.10.2"
typer = "^0.12.3"
scikit-learn = "^1.5.1"
pykdtree = "^1.3.12"
kmedoids = "^0.5.1"
colour-science = "^0.4.4"
rich = "^13.7.1"
matplotlib = "^3.9.0"
psutil = "^6.0.0"
pymixbox = "^2.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
